Remember the envelope sender and recipient, remember the connecting IP, and insert appropriate headers at the top of the message we send to spamc, so that Spamassassin can use them.
The envelope info gets put into X-Envelope-From: and X-Envelope-To: ,
which SA will use when checking the whitelist_to and _from rules.

The connecting IP gets put into a dummy "Received: [111.222.333.444]"
header.  Without it, none of the DNSBL or whitelist checks will see
this IP.  The dummy header does not seem to trigger any "forged
Received:" rules, so that's good.

Note that remembering the connecting IP address means we need to carry
information between the lifetimes of SpamAssassin objects, so we have
to use the milter private data pointer.  This means that within a
message, the pointer points to a SpamAssassin object.  Outside of a
message but during a client connection, the pointer points to a struct
connect_info.  This means some extra work in the message error path to
recover the connect_info pointer.  It also means that for symmetry, we
have to create the SpamAssassin object at the beginning on a "message"
session - i.e. within mlfi_envrcpt instead of mlfi_header.

Also for ease of coding add some more output methods with string and
unsized void* arguments.